dashboard/backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
# AIND_SIMILARITY 적응 패키지
python-Levenshtein>=0.25.0
nltk>=3.9.0
scikit-learn>=1.5.0
